1.增加删除 2.增加自动恢复数据

+++ lawoffice-server/src/main/java/com/julius/law/project/mapper/ProjectconcatMapper.java
@@ -25,5 +25,5 @@ public interface ProjectconcatMapper extends BaseMapper<Projectconcat> {
      * @param projectId
      */
     void insertBatch(@Param("projectConcats") List<Projectconcat> projectConcats,
-                     @Param("projectId") Integer projectId);
+                     @Param("projectId") Long projectId);
 }
